+++ web/app/components/base/amplitude/index.ts
@@ -1,2 +1,3 @@
-export { default, isAmplitudeEnabled } from './AmplitudeProvider'
+export { isAmplitudeEnabled } from './AmplitudeProvider'
+export { default } from './lazy-amplitude-provider'
 export { resetUser, setUserId, setUserProperties, trackEvent } from './utils'
