+++ web/app/components/sentry-initializer.tsx
@@ -1,23 +1,35 @@
 'use client'
 
 import type { ReactElement } from 'react'
+import * as Sentry from '@sentry/react'
+import { useEffect } from 'react'
 import { IS_DEV } from '@/config'
 import { env } from '@/env'
-import dynamic from '@/next/dynamic'
-
-const SentrySetup = dynamic(() => import('./sentry-setup'), { ssr: false })
 
 const SentryInitializer = ({
   children,
 }: { children: ReactElement }) => {
-  const SENTRY_DSN = env.NEXT_PUBLIC_SENTRY_DSN
-
-  return (
-    <>
-      {children}
-      {!IS_DEV && SENTRY_DSN && <SentrySetup dsn={SENTRY_DSN} />}
-    </>
-  )
+  useEffect(() => {
+    const SENTRY_DSN = env.NEXT_PUBLIC_SENTRY_DSN
+    if (!IS_DEV && SENTRY_DSN) {
+      try {
+        Sentry.init({
+          dsn: SENTRY_DSN,
+          integrations: [
+            Sentry.browserTracingIntegration(),
+            Sentry.replayIntegration(),
+          ],
+          tracesSampleRate: 0.1,
+          replaysSessionSampleRate: 0.1,
+          replaysOnErrorSampleRate: 1.0,
+        })
+      }
+      catch (error) {
+        console.error('Failed to initialize Sentry', error)
+      }
+    }
+  }, [])
+  return children
 }
 
 export default SentryInitializer
